Edit file structure
