+++ .gitignore
@@ -1,0 +1,13 @@
+# Dependency directories
+node_modules
+
+# Logs
+logs
+*.log
+
+# dotenv environment variable files
+env
+.env
+
+#IDEA
+.idea
