+++ README.md
@@ -1,1 +1,7 @@
 # FarmUs_Node
+
+<br />
+
+```cmd
+nodemon start
+```
