+++ components/user/userRoute.js
@@ -1,0 +1,7 @@
+module.exports = function(app){
+    const user = require('./userController');
+    //const jwtMiddleware = require('../../../config/jwtMiddleware');
+
+    // 0. 테스트 API
+    app.get('/app/test', user.getTest)
+};
