+++ config/cors.js
@@ -1,0 +1,4 @@
+module.exports = {
+        origin: "http://localhost:3000",
+        credentials: true,
+}
