+++ config/response.js
@@ -1,0 +1,18 @@
+const response = ({isSuccess, code, message}, result) => {
+    return {
+         isSuccess: isSuccess,
+         code: code,
+         message: message,
+         result: result
+    }
+   };
+
+   const errResponse = ({isSuccess, code, message}) => {
+     return {
+         isSuccess: isSuccess,
+         code: code,
+         message: message
+       }
+   };
+
+   module.exports = { response, errResponse };
