--- farmus/components/user/userDao.js
@@ -1,1 +1,0 @@
-//sql문
