--- farmus/components/user/userService.js
@@ -1,2 +1,0 @@
-// Service: Create, Update, Delete 비즈니스 로직 처리
-
